+++ config.js
@@ -21,5 +21,5 @@
 window.CARDIO_CONFIG = {
   syncUrl: '',
   syncToken: '',
-  appVersion: '1.2.1'
+  appVersion: '1.3.0'
 };